+++ mitra_activitypub/src/handlers/add.rs
@@ -111,11 +111,7 @@ async fn handle_fep_171b_add(
                 },
             };
         },
-        Err(AuthenticationError::DatabaseError(db_error)) => return Err(db_error.into()),
-        Err(other_error) => {
-            log::warn!("{other_error}");
-            return Err(ValidationError("invalid integrity proof").into());
-        },
+        Err(other_error) => return Err(other_error.into()),
     };
     // Authorization
     if !is_same_origin(&conversation_owner, &target)? {